drivers: soc: qcom: pil-msa: Use PIL device for MBA DMA allocations
The idea of using generic allocations is absolutely valid and
I absolutely agree, but this is making MBA not to load, choosing
the wrong physical address where to load the MBA OS.
This makes the PBL to succeed, but, since it is loaded to a different
(yet valid) address, the MBA OS will panic because of that.

The alternative solution is to calculate the address to relocate
the firmware, but that would actually need definitely more code,
even if it's A LOT cleaner than this.

This commit will make us able to use the old allocation way for MBA,
letting it stay between the modem's CMA region, only if the loaded
image is not in ELF format (*.p00), as it is an old one, failing with
the addressing change.

Actually strictly needed for MSM8974 and MSM8226 devices.