drivers: vidc: Set operating rate to fps if not set by userspace
The default NuPlayer in android framework sets the operating rate as :-

operating rate = fps * playback_speed

But some players like HW+ codec in MXPlayer and WebRTC in chrome do not
set it. As a result, if a variable fps video is played in these players
whose frame rate drops to 0 fps at some point, this leads to venus HW
moving to sleep state due to clock scaling which results in hung HW.
So, set the operating rate same as fps assuming playback speed to be 1.0
(Atleast its better than hung HW) incase not set by the userspace.
The real issue maybe somewhere down under but for the time being,
this fixes the issue for me.